Try to disable wifi powersave via iwconfig
Some drivers are not yet converted to cfg80211 and don't support nl80211
(required by `iw`). These drivers are usually using `iwconfig` and `iwpriv`
to adjust various settings. Just trying to run `iwconfig .. power off` to
disable powersave can therefore be used (sometimes) to disable the
powersave feature on legacy drivers. And since the return code of this
command is ignored, it will not stop the wireless.js service in case it
fails.

Signed-off-by: Sven Eckelmann <sven@narfation.org>