Fix wireless power on wext-only drivers (#79)
* Make disabling of wifi powersave optional

Not all wifi drivers have been switched from wext to nl80211 (used by
`iw`). And not all nl80211 compatible wifi drivers support the power_save
configuration. On these kind of devices, the service would fail due to the
exit code of the command in ExecStartPost. And since it wasn't an optional
(ignoring exit code), the wireless.js daemon would also be stopped by
systemd. As result, the wireless would work for some time but at some point
suddenly stop working (e.g. after 24h).

Since the disabling of powersave is not mission-critical, making it an
optional command (which can fail without stopping the service) is the more
desirable outcome.

Fixes: 76060f11008d ("Relocate `power_save` to wireless.service and remove `wifistart.sh`")
Signed-off-by: Sven Eckelmann <sven@narfation.org>

* Try to disable wifi powersave via iwconfig

Some drivers are not yet converted to cfg80211 and don't support nl80211
(required by `iw`). These drivers are usually using `iwconfig` and `iwpriv`
to adjust various settings. Just trying to run `iwconfig .. power off` to
disable powersave can therefore be used (sometimes) to disable the
powersave feature on legacy drivers. And since the return code of this
command is ignored, it will not stop the wireless.js service in case it
fails.

Signed-off-by: Sven Eckelmann <sven@narfation.org>

---------

Signed-off-by: Sven Eckelmann <sven@narfation.org>